backend/requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.8
django-redis==5.4.0
kafka-python==2.0.2
gunicorn==22.0.0
whitenoise==6.7.0
python-decouple==3.8
Pillow==10.4.0
